Move queries from controller to model
